requirements.txt
streamlit
pandas
scikit-learn
joblib
numpy
faiss-cpu
sentence-transformers
groq